authproject/authproject/requirements.txt
djangorestframework
django
djangorestframework_simplejwt